+++ grpc_host/lib/src/configuration/configuration.dart
@@ -1,5 +1,4 @@
 export 'file_yaml_extension.dart';
 export 'host_settings.dart';
-export 'smtp_settings.dart';
 export 'token_settings.dart';
 export 'settings.dart';