+++ grpc_host/lib/src/hosting/principal_interceptor.dart
@@ -20,5 +20,6 @@ abstract class PrincipalInterceptor {
       getTokenPayload: getTokenPayload,
       createPrincipal: createPrincipal,
     );
+    return null;
   }
 }